requirements.txt
altgraph==0.17.5
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-WTF==1.2.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
packaging==26.0
pefile==2024.8.26
psycopg2-binary==2.9.11
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
Werkzeug==3.1.5
WTForms==3.2.1
